RFCT(Further changes to mocked tests to avoid the calling of Win32 libraries)
