+++ src/main/java/mmarquee/automation/controls/AutomationToolBar.java
@@ -18,6 +18,7 @@ package mmarquee.automation.controls;
 import mmarquee.automation.AutomationElement;
 import mmarquee.automation.AutomationException;
 import mmarquee.automation.ControlType;
+import mmarquee.automation.UIAutomation;
 import mmarquee.automation.pattern.ItemContainer;
 import mmarquee.automation.pattern.PatternNotFoundException;
 
@@ -49,6 +50,19 @@ public class AutomationToolBar extends AutomationContainer {
      */
     public AutomationToolBar(AutomationElement element, ItemContainer container) throws AutomationException, PatternNotFoundException {
         super(element, container);
+    }
+
+    /**
+     * Constructor for the AutomationToolBar.
+     *
+     * @param element The underlying automation element
+     * @param container The ItemContainer pattern
+     * @param instance The automation instance
+     * @throws AutomationException Automation library error
+     * @throws PatternNotFoundException Could not find pattern
+     */
+    public AutomationToolBar(AutomationElement element, ItemContainer container, UIAutomation instance) throws AutomationException, PatternNotFoundException {
+        super(element, container, instance);
     }
 
     /**
