+++ src/test/java/mmarquee/automation/UIAutomationTest.java
@@ -98,6 +98,13 @@ public class UIAutomationTest extends BaseAutomationTest {
     }
 
     @Test
+    public void testCreateAutomationIdPropertyCondition_Does_Not_Throw_Exception() throws AutomationException {
+        UIAutomation instance = UIAutomation.getInstance();
+
+        instance.createAutomationIdPropertyCondition("ID");
+    }
+
+    @Test
     public void testCompareElementsWhenTheSame() {
         UIAutomation instance = UIAutomation.getInstance();
 
@@ -349,12 +356,7 @@ public class UIAutomationTest extends BaseAutomationTest {
         instance.createNamePropertyCondition("ID");
     }
 
-    @Test
-    public void testCreateAutomationIdPropertyCondition_Does_Not_Throw_Exception() throws AutomationException {
-        UIAutomation instance = UIAutomation.getInstance();
 
-        instance.createAutomationIdPropertyCondition("ID");
-    }
 
     @Test
     public void testCreateControlTypeCondition_Does_Not_Throw_Exception() throws AutomationException {
@@ -459,18 +461,6 @@ public class UIAutomationTest extends BaseAutomationTest {
     }
 
     @Test(expected = AutomationException.class)
-    public void testCreateFalseCondition_Throws_Exception_When_Automation_Returns_False()
-            throws AutomationException {
-        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
-
-        when(mocked_automation.createFalseCondition(isA(PointerByReference.class))).thenReturn(-1);
-
-        UIAutomation local_instance = new UIAutomation(mocked_automation);
-
-        local_instance.createFalseCondition();
-    }
-
-    @Test(expected = AutomationException.class)
     @Ignore("Mocking fails - needs investigation")
     public void testCreateAndCondition_Throws_Exception_When_Automation_Returns_False()
             throws AutomationException {
@@ -500,29 +490,6 @@ public class UIAutomationTest extends BaseAutomationTest {
     public void test_GetDesktopMenu_Throws_Exception_When_Not_Found()
             throws IOException, AutomationException, PatternNotFoundException {
         UIAutomation.getInstance().getDesktopMenu("Menu");
-    }
-
-    @Test
-    public void testGetDesktopWindow_Succeeds_When_Window_Present() throws IOException, AutomationException, PatternNotFoundException {
-        IUIAutomation mocked = Mockito.mock(IUIAutomation.class);
-
-        doAnswer(new Answer() {
-            @Override
-            public WinNT.HRESULT answer(InvocationOnMock invocation) throws Throwable {
-                return new WinNT.HRESULT(0);
-            }
-        }).when(mockUnknown).QueryInterface(any(), any());
-
-        when(mocked.createAndCondition(any(Pointer.class), any(Pointer.class), any(PointerByReference.class))).thenReturn(0);
-        when(mocked.createPropertyCondition(any(Integer.class), any(Variant.VARIANT.ByValue.class), any(PointerByReference.class))).thenReturn(0);
-
-        UIAutomation local_instance = Mockito.mock(UIAutomation.class);
-
-        doReturn(mockUnknown)
-                .when(local_instance)
-                .makeUnknown(any(Pointer.class));
-
-        local_instance.getDesktopWindow(getLocal("notepad.title"));
     }
 
     @Test(expected = ItemNotFoundException.class)
