+++ src/test/java/mmarquee/automation/controls/AutomationToolbarTest.java
@@ -16,7 +16,9 @@
 package mmarquee.automation.controls;
 
 import mmarquee.automation.AutomationElement;
+import mmarquee.automation.UIAutomation;
 import mmarquee.automation.pattern.ItemContainer;
+import mmarquee.automation.uiautomation.IUIAutomation;
 import mmarquee.automation.uiautomation.IUIAutomationElement3;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -43,7 +45,11 @@ public class AutomationToolbarTest {
         ItemContainer container = Mockito.mock(ItemContainer.class);
         when(element.getName()).thenReturn("NAME");
 
-        AutomationToolBar ctrl = new AutomationToolBar(element, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationToolBar ctrl = new AutomationToolBar(element, container, instance);
 
         String name = ctrl.getName();
 
@@ -62,7 +68,11 @@ public class AutomationToolbarTest {
 
         when(element.findAll(any(), any())).thenReturn(result);
 
-        AutomationToolBar ctrl = new AutomationToolBar(element, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationToolBar ctrl = new AutomationToolBar(element, container, instance);
 
         AutomationToolBarButton btn = ctrl.getToolbarButton(0);
     }
@@ -79,7 +89,11 @@ public class AutomationToolbarTest {
 
         when(element.findAll(any(), any())).thenReturn(result);
 
-        AutomationToolBar ctrl = new AutomationToolBar(element, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationToolBar ctrl = new AutomationToolBar(element, container, instance);
 
         AutomationToolBarButton btn = ctrl.getToolbarButton(1);
     }
